make Command interface not sealed
